// src/apis/models/GameSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pick Em League Swagger
 * Welcome to the Swagger UI documentation site!
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { TeamSchema } from './TeamSchema';
import {
    TeamSchemaFromJSON,
    TeamSchemaFromJSONTyped,
    TeamSchemaToJSON,
} from './TeamSchema';

/**
 * 
 * @export
 * @interface GameSchema
 */
export interface GameSchema {
    /**
     * 
     * @type {number}
     * @memberof GameSchema
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof GameSchema
     */
    gameTime?: string;
    /**
     * 
     * @type {number}
     * @memberof GameSchema
     */
    week?: number;
    /**
     * 
     * @type {number}
     * @memberof GameSchema
     */
    result?: number;
    /**
     * 
     * @type {TeamSchema}
     * @memberof GameSchema
     */
    homeTeam?: TeamSchema;
    /**
     * 
     * @type {TeamSchema}
     * @memberof GameSchema
     */
    awayTeam?: TeamSchema;
}

/**
 * Check if a given object implements the GameSchema interface.
 */
export function instanceOfGameSchema(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GameSchemaFromJSON(json: any): GameSchema {
    return GameSchemaFromJSONTyped(json, false);
}

export function GameSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameSchema {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'gameTime': !exists(json, 'game_time') ? undefined : json['game_time'],
        'week': !exists(json, 'week') ? undefined : json['week'],
        'result': !exists(json, 'result') ? undefined : json['result'],
        'homeTeam': !exists(json, 'home_team') ? undefined : TeamSchemaFromJSON(json['home_team']),
        'awayTeam': !exists(json, 'away_team') ? undefined : TeamSchemaFromJSON(json['away_team']),
    };
}

export function GameSchemaToJSON(value?: GameSchema | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'game_time': value.gameTime,
        'week': value.week,
        'result': value.result,
        'home_team': TeamSchemaToJSON(value.homeTeam),
        'away_team': TeamSchemaToJSON(value.awayTeam),
    };
}